+++ OpenNos.Data/CharacterDTO.cs
@@ -39,9 +39,9 @@ namespace OpenNos.Data
         public long Gold { get; set; }
         public int Backpack { get; set; }
         public byte Level { get; set; }
-        public int LevelXp { get; set; }
+        public long LevelXp { get; set; }
         public byte JobLevel { get; set; }
-        public int JobLevelXp { get; set; }
+        public long JobLevelXp { get; set; }
         public int Act4Dead { get; set; }
         public int Act4Kill { get; set; }
         public int Faction { get; set; }
